Upgrade to react 0.14.0
Summary:
This modifies the build script to point to v0.14.0 and fixes up
some paths and warnings for the new React. This also exports ReactDOM
attached to the React object, so webapp has access to it without there
being a separate file.

I also added some custom warnings when using React.addons stuff, because
the "don't require('react/addons')" error is silenced (since it is my
fault that everyone will see it, not the fault of someone requiring
react).

Test Plan:
 - `make clean; make; make install WEBAPP=../webapp`
 - Visit http://localhost:8080/
 - See that we get appropriate warnings.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: https://phabricator.khanacademy.org/D22081