Add react-dom/server to the React bundle.
Summary:
Perseus requires react-dom/server for some of it's tests. This adds in
react-dom/server to the main react bundle.

I tested the size impact this has, and it looks like it's minimal (i.e. ~200
bytes after minification and gzip). I'd be willing to bet that the code was
already included, this is just adding a place where it's exposed to the world.

Test Plan: - `make`

Reviewers: john

Reviewed By: john

Subscribers: jordan, lauren

Differential Revision: https://phabricator.khanacademy.org/D33598