Removes React Art and updates version to React 16.3.0
Summary:
We don't use React Art in webapp anymore, so we can remove it from the React bundling process.

In addition, the addons that we’ve used up until now have been deprecated with React 15.5. (https://reactjs.org/docs/addons.html) As a result, we require and export more individual libraries with React 16, and must update the corresponding files in javascript/node_modules where appropriate. This adds many of those libraries to the React object so we can reference them properly.

Test Plan:
* `make clean`
* `make`
* `make install WEBAPP=../webapp`
* See that the correct version is visible in webapp

Reviewers: emily, scottgrant

Reviewed By: emily

Subscribers: michaelpolyak, tom, colin

Differential Revision: https://phabricator.khanacademy.org/D43642